refactor: move VIP to its own folder, following the new layout
